Add stripe map; move mock superstripe helpers to jwst/lib/tests
