+++ jwst/dq_init/tests/helpers.py
@@ -1,85 +1,12 @@
 import numpy as np
-from stdatamodels.jwst.datamodels import (
-    GuiderRawModel,
-    MaskModel,
-    RampModel,
-    SuperstripeRampModel,
-    dqflags,
-)
+from stdatamodels.jwst.datamodels import MaskModel
+
+from jwst.lib.tests.helpers import make_rawramp
 
 __all__ = [
-    "make_rawramp",
     "make_rampmodel",
     "make_dq_arrays",
-    "make_superstripe_model",
-    "one_stripe_mask",
-    "make_superstripe_mask_model",
 ]
-
-
-def make_rawramp(
-    instrument, nints, ngroups, ysize, xsize, ystart, xstart, exp_type=None, num_superstripe=None
-):
-    """
-    Make a raw ramp model for any instrument.
-
-    Parameters
-    ----------
-    instrument : str
-        Instrument name. If "FGS", a ``GuiderRawModel`` is returned.
-    nints : int
-        Number of integrations.
-    ngroups : int
-        Number of groups.
-    ysize : int
-        Y-size for the subarray.
-    xsize : int
-        X-size for the subarray.
-    ystart : int
-        Y-start value for the subarray.
-    xstart : int
-        X-start value for the subarray.
-    exp_type : str, optional
-        Exposure type.
-    num_superstripe : int, optional
-        Number of superstripes. If not None, a ``SuperstripeRampModel``
-        is returned.
-
-    Returns
-    -------
-    dm_ramp : `~stdatamodels.jwst.datamodels.RampModel`, \
-              `~stdatamodels.jwst.datamodels.GuiderRawModel`, or \
-              `~stdatamodels.jwst.datamodels.SuperstripeRampModel`
-        The ramp model.
-    """
-    # create the data and groupdq arrays
-    csize = (nints, ngroups, ysize, xsize)
-    data = np.full(csize, 1.0)
-
-    # create a JWST datamodel
-    if instrument == "FGS":
-        dm_ramp = GuiderRawModel(data=data)
-    elif num_superstripe is not None:
-        dm_ramp = SuperstripeRampModel(data=data)
-        dm_ramp.meta.subarray.num_superstripe = num_superstripe
-    else:
-        dm_ramp = RampModel(data=data)
-
-    dm_ramp.meta.exposure.type = exp_type
-    dm_ramp.meta.subarray.xstart = xstart
-    dm_ramp.meta.subarray.xsize = xsize
-    dm_ramp.meta.subarray.ystart = ystart
-    dm_ramp.meta.subarray.ysize = ysize
-
-    # add some basic metadata
-    dm_ramp.meta.instrument.name = instrument
-    dm_ramp.meta.observation.date = "2026-01-01"
-    dm_ramp.meta.observation.time = "00:00:00"
-    dm_ramp.meta.exposure.nints = nints
-    dm_ramp.meta.exposure.ngroups = ngroups
-    dm_ramp.meta.exposure.nframes = 1
-
-    return dm_ramp
 
 
 def make_rampmodel(nints=1, ngroups=5, ysize=1024, xsize=1032):
@@ -147,173 +74,3 @@ def make_dq_arrays(ysize, xsize):
     dq_def = np.array((dqdef), dtype=mask.get_dtype("dq_def"))
 
     return dq, dq_def
-
-
-def int_times_table(n_entry):
-    """
-    Make an integration times table, suitable for storing as ``model.int_times``.
-
-    Parameters
-    ----------
-    n_entry : int
-        Length of the integration table to create
-
-    Returns
-    -------
-    int_times : ndarray
-        Record array with expected time fields. Integrations are regularly
-        spaced, with a little noise in the recorded time.
-    """
-    int1_times = [
-        60897.76391539,
-        60897.76393129,
-        60897.76394719,
-        60897.76855646,
-        60897.76857236,
-        60897.76858826,
-    ]
-    integrations = []
-    rng = np.random.default_rng(seed=42)
-    for i in range(n_entry):
-        # add new integrations with regular spacing, plus a little noise for non-trivial tests
-        new_int = tuple([i + 1] + [t + 3.2e-5 * i + 1e-7 * rng.random() for t in int1_times])
-        integrations.append(new_int)
-
-    integration_table = np.array(
-        integrations,
-        dtype=[
-            ("integration_number", "i4"),
-            ("int_start_MJD_UTC", "f8"),
-            ("int_mid_MJD_UTC", "f8"),
-            ("int_end_MJD_UTC", "f8"),
-            ("int_start_BJD_TDB", "f8"),
-            ("int_mid_BJD_TDB", "f8"),
-            ("int_end_BJD_TDB", "f8"),
-        ],
-    )
-    return integration_table
-
-
-def make_superstripe_model(add_inttimes=False, add_zeroframe=False):
-    """
-    Make a NIRISS SOSS superstripe raw ramp model.
-
-    Parameters
-    ----------
-    add_inttimes : bool, optional
-        If True, an `int_times` table will be attached to the output datamodel.
-    add_zeroframe : bool, optional
-        If True, a `zeroframe` array will be attached to the output datamodel.
-        In real data, the zeroframe is only used for NIRCam. It is attached to
-        this NIRISS model for testing convenience.
-
-    Returns
-    -------
-    `~stdatamodels.jwst.datamodels.RampModel`
-        The ramp model.
-    """
-    nints = 3
-    nstripe = 10
-    ngroups = 5
-    ysize = 256
-    xsize = 208
-    xstart = 1
-    ystart = 1793
-    model = make_rawramp(
-        "NIRISS", nints * nstripe, ngroups, ysize, xsize, ystart, xstart, num_superstripe=nstripe
-    )
-
-    # Add some more basic metadata
-    model.meta.instrument.detector = "NIS"
-    model.meta.subarray.name = "SUB204STRIPE_SOSS"
-    model.meta.subarray.fastaxis = -2
-    model.meta.subarray.slowaxis = -1
-    model.meta.exposure.integration_start = 1
-    model.meta.exposure.integration_end = nints * nstripe
-    model.meta.exposure.nints = nints * nstripe
-
-    # Add the superstripe metadata
-    model.meta.subarray.multistripe_reads1 = 4
-    model.meta.subarray.multistripe_reads2 = 204
-    model.meta.subarray.multistripe_skips1 = 0
-    model.meta.subarray.multistripe_skips2 = 0
-    model.meta.subarray.num_superstripe = nstripe
-    model.meta.subarray.repeat_stripe = 1
-    model.meta.subarray.interleave_reads1 = 1
-    model.meta.subarray.superstripe_step = 204
-
-    if add_inttimes:
-        model.int_times = int_times_table(nints * nstripe)
-
-    if add_zeroframe:
-        model.meta.exposure.zero_frame = True
-        model.zeroframe = model.data[:, 0, :, :].copy()
-
-    return model
-
-
-def one_stripe_mask(shape):
-    """
-    Make a DQ mask with a range of bad pixels for a single stripe.
-
-    Parameters
-    ----------
-    shape : tuple of int
-        Stripe data shape.
-
-    Returns
-    -------
-    dq : ndarray
-        DQ array matching the stripe with bad pixels.
-    """
-    dq = np.zeros(shape, dtype=np.uint32)
-    dq[2, 100] = dqflags.pixel["DEAD"]
-    dq[4, 100] = dqflags.pixel["HOT"]
-    dq[6, 100] = dqflags.pixel["UNRELIABLE_SLOPE"]
-    dq[8, 100] = dqflags.pixel["RC"]
-    dq[10, 100] = dqflags.pixel["DO_NOT_USE"]
-    dq[2, 200] = dqflags.pixel["DO_NOT_USE"] + dqflags.pixel["DEAD"]
-    dq[4, 200] = dqflags.pixel["DO_NOT_USE"] + dqflags.pixel["HOT"]
-    dq[6, 200] = dqflags.pixel["DO_NOT_USE"] + dqflags.pixel["UNRELIABLE_SLOPE"]
-    dq[8, 200] = dqflags.pixel["DO_NOT_USE"] + dqflags.pixel["RC"]
-    return dq
-
-
-def make_superstripe_mask_model():
-    """
-    Create a MaskModel with bad pixels in a superstripe region.
-
-    Returns
-    -------
-    `~stdatamodels.jwst.datamodels.MaskModel`
-        The mask model.
-    """
-    ysize = 2048
-    xsize = 2048
-    dq = np.zeros((ysize, xsize), dtype=np.uint32)
-
-    # edit reference file with known bad pixel values
-    substart = 1792
-    subsize = 256
-    nstripe = 10
-    stripe = 204
-    stripe_shape = (subsize, stripe)
-    for i in range(nstripe):
-        xstart = stripe * i + 4
-        xstop = xstart + stripe
-        dq[substart : substart + subsize, xstart:xstop] = one_stripe_mask(stripe_shape)
-
-    # write mask model
-    ref_data = MaskModel(dq=dq)
-    ref_data.meta.instrument.name = "NIRISS"
-    ref_data.meta.subarray.xstart = 1
-    ref_data.meta.subarray.xsize = xsize
-    ref_data.meta.subarray.ystart = 1
-    ref_data.meta.subarray.ysize = ysize
-    ref_data.meta.description = "Test description"
-    ref_data.meta.reftype = "mask"
-    ref_data.meta.author = "Test Author"
-    ref_data.meta.pedigree = "test"
-    ref_data.meta.useafter = "2024-01-01"
-
-    return ref_data
